+++ HISTORY.md
@@ -3,9 +3,10 @@
 ## in development:
 
 * **New Features:**
+    * support deleting and inserting individual variadic slots, script vars & ring params
     * support dragging blocks, costumes and sounds out from result bubbles, and from speech balloons and variable watchers when in edit mode
     * support exporting costumes and sounds from result bubbles, and from speech balloons and variable watchers when in edit mode
-    * support deleting and inserting individual variadic slots, script vars & ring params
+    * support exporting text and numbers from result bubbles
 * **Notable Changes:**
 * **Notable Fixes:**
     * fixed layout for scrolling custom categories, thanks, Eckart, for the bug report!
@@ -21,6 +22,7 @@
 * blocks: support exporting sounds from result bubbles
 * objects: support exporting sounds and costumes from speech balloons in edit mode
 * objects: support exporting sounds and costumes from variable watchers in edit mode
+* blocks: support exporting numbers and text from result bubbles
 
 ### 2022-01-27
 * blocks, gui: support dragging costumes and sounds out from result bubbles